lab 4 remove todo code
