+++ quests/dataflow/4_SQL_Batch_Analytics/labs/pom.xml
@@ -21,15 +21,15 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.mypackage</groupId>
-    <artifactId>advanced-dataflow-lab-3-sql-solution</artifactId>
+    <artifactId>advanced-dataflow-lab-4-lab</artifactId>
     <version>1.0</version>
 
     <properties>
         <!-- Dependency properties -->
-        <beam.version>2.23.0</beam.version>
+        <beam.version>2.22.0</beam.version>
         <java.version>1.8</java.version>
         <slf4j.version>1.7.25</slf4j.version>
-        <maven-shade-plugin.version>3.2.3</maven-shade-plugin.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven-dependency-plugin.version>3.0.1</maven-dependency-plugin.version>
     </properties>
 
@@ -59,16 +59,7 @@
             <artifactId>beam-sdks-java-extensions-google-cloud-platform-core</artifactId>
             <version>${beam.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.apache.beam</groupId>
-            <artifactId>beam-sdks-java-extensions-sql</artifactId>
-            <version>${beam.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.beam</groupId>
-            <artifactId>beam-sdks-java-extensions-sql-zetasql</artifactId>
-            <version>${beam.version}</version>
-        </dependency>
+
         <!-- Logging -->
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -92,39 +83,6 @@
                     <source>${java.version}</source>
                     <target>${java.version}</target>
                 </configuration>
-            </plugin>
-            <!-- The maven shade plugin is used to create an uber-jar with all the
-              dependencies needed to run as a standalone jar.
-              Do not minimize the jar since that removes some of the required
-              classes for the runners. -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>${maven-shade-plugin.version}</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                        <configuration>
-                            <transformers>
-                                <transformer
-                                        implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
-                            </transformers>
-                            <filters>
-                                <filter>
-                                    <artifact>*:*</artifact>
-                                    <excludes>
-                                        <exclude>META-INF/*.SF</exclude>
-                                        <exclude>META-INF/*.DSA</exclude>
-                                        <exclude>META-INF/*.RSA</exclude>
-                                    </excludes>
-                                </filter>
-                            </filters>
-                        </configuration>
-                    </execution>
-                </executions>
             </plugin>
         </plugins>
     </build>
