--- quests/dataflow/4_SQL_Batch_Analytics/solution_bak/README.md
@@ -1,23 +1,0 @@
-# Project Name
-
-Project Description
-
-
-## Pipeline
-
-[SamplePipeline](src/main/java/com/google/cloud/pso/pipeline/SamplePipeline.java) -
-A sample pipeline which can be used as a base for creating a real Dataflow pipeline.
-
-## Getting Started
-
-### Requirements
-
-* Java 8
-* Maven 3
-
-### Building the Project
-
-Build the entire project using the maven compile command.
-```sh
-mvn clean && mvn compile
-```
