--- quests/dataflow/4_SQL_Batch_Analytics/solution_bak/src/license/license-header-style.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<additionalHeaders>
-  <slashstar_style>
-    <firstLine>/*</firstLine>
-    <beforeEachLine> * </beforeEachLine>
-    <endLine> */&#10;</endLine>
-    <firstLineDetectionPattern>(\s|\t)*/\*.*$</firstLineDetectionPattern>
-    <lastLineDetectionPattern>.*\*/(\s|\t)*$</lastLineDetectionPattern>
-    <allowBlankLines>false</allowBlankLines>
-    <isMultiline>true</isMultiline>
-    <padLines>false</padLines>
-  </slashstar_style>
-</additionalHeaders>
