--- quests/dataflow/4_SQL_Batch_Analytics/solution_bak/src/main/resources/spotbugs-exclude.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2018 Google Inc.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<!--
-  This file contains some false positive bugs detected by findbugs. Their
-  false positive nature has been analyzed individually and they have been
-  put here to instruct findbugs it must ignore them.
-
-  For more information on adding filters to this file, refer to:
-  https://spotbugs.readthedocs.io/en/stable/filter.html
--->
-<FindBugsFilter>
-  <!-- Ignored bug categories. Bugs in these categories should not block the build. -->
-  <Bug category="I18N"/>
-  <Bug pattern="DM_STRING_CTOR"/>
-  <Bug pattern="EI_EXPOSE_REP" />
-  <Bug pattern="EI_EXPOSE_REP2" />
-  <Bug pattern="SE_NO_SERIALVERSIONID"/>
-
-  <!-- Suppress checking of AutoValue internals -->
-  <Match>
-    <Class name="~.*AutoValue_.*"/>
-  </Match>
-
-  <!--
-    Suppressed findbugs issues. All new issues should include a comment why they're
-    suppressed. Suppressions should go in this file rather than inline using
-    @SuppressFBWarnings to avoid unapproved artifact license.
-	-->
-
-</FindBugsFilter>
